requirements.txt
opencv-contrib-python==4.5.5.64
numpy==1.22.3
matplotlib~=3.5.1
metawear~=1.0.6
pywin32~=303;platform_system=="Windows"
python3-Xlib;platform_system=="Linux"
